requirements.txt
django==4.2
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
celery==5.3.6
psycopg2==2.9.9